调整了刷新埋点配置的时机，默认5分钟刷新一次

+++ library/src/main/java/io/sugo/android/mpmetrics/SGConfig.java
@@ -206,7 +206,7 @@ public class SGConfig {
 
         mBulkUploadLimit = metaData.getInt("io.sugo.android.SGConfig.BulkUploadLimit", 40); // 40 records default
         mFlushInterval = metaData.getInt("io.sugo.android.SGConfig.FlushInterval", 60 * 1000); // one minute default
-        mUpdateDecideInterval = metaData.getLong("io.sugo.android.SGConfig.UpdateDecideInterval", 5 * 60 * 1000); // 5 minute default
+        mUpdateDecideInterval = metaData.getInt("io.sugo.android.SGConfig.UpdateDecideInterval", 5 * 60 * 1000); // 5 minute default
         mDataExpiration = metaData.getInt("io.sugo.android.SGConfig.DataExpiration", 1000 * 60 * 60 * 24 * 5); // 5 days default
         mMinimumDatabaseLimit = metaData.getInt("io.sugo.android.SGConfig.MinimumDatabaseLimit", 20 * 1024 * 1024); // 20 Mb
         mDisableFallback = metaData.getBoolean("io.sugo.android.SGConfig.DisableFallback", true);
